lib: fix extract_event_data State::Sub SE processing
Only emit `SubNegotiation` type event when in `State::Sub` and `IAC SE`
is encountered in sequence, not just `SE`.

When extracting event data, if the parser is in the `State::Sub` state,
it needs to look for the end of the sub-negotiation data to decide
whether to push a `EventType::SubNegotiation` event.

Prior to this branch `extract_event_data` did so unconditionally upon
encountering a `SE` value (0xF0) anytime after entering the `Sub` state.
This incorrectly produces a `SubNegotiation` event type even if the `SE`
value wasn't preceded by an `IAC` value (0xFF). This in turn leads to
data being dropped, and the subnegotation data being turned into
a truncated application data read.

For example, receiving the following data:

```
  [IAC, SB, GMCP, SE, 0x9F, 0x91, 0x8B, IAC, SE];
```

The `extract_event_data` fn would produce a `EventType::SubNegotiation`
event type with the buf

```
  [IAC, SB, GMCP, SE]
```

and a `EventType::None` event type with the buf

```
  [0x9F, 0x91, 0x8b]
```

Later, in `process`, these two event types are processed.

First the `EventType::SubNegotiation` match arm will find the event
buffer doesn't end with `IAC SE`, and so pushes the event buffer into
its processing buffer, making `self.buffer` hold:

```
  [IAC, SB, GMCP, SE]
```

Processing continues to the next event, and the `EventType::None` match
arm. This gets processed as a `TelnetEvents::DataReceive`, and the event
buffer is passed as the application data.

We then reach the end of the events, and so the `self.buffer` content is
cleared, and only the `TelnetEvents::DataReceive` event is returned,
with the truncated data:

```
  [0x9F, 0x91, 0x8b]
```

The result _should_ be a `TelnetEvents::SubNegotiation` with `opt ==
GMCP` and `buffer == [SE, 0x9F, 0x91, 0x8B]`.

For that to happen `extract_event_data` needs to be changed to only emit
the `SubNegotiation` event when it encounters the SE _after_ an IAC.
This presents the entire subnegotiation event type buffer in tact to
`process`, where it is turned into the expected
`TelnetEvents::Subnegotiation` return.

This commit implements that fix, and adds a unit test matching the
scenario described above.