lib: inline dbuffer binding
Both 'buffer' and 'dbuffer' names were in-scope in this state handling
code. The 'dbuffer' name isn't very descriptive - it's the buffer
holding the subnegotiation parameters found between the command
start/end markers.

Since the variable is only used in one place it's easier to inline the
expression than rename the variable for clarity.