lib: add SubIac state to extract_event_data
Rather than looking backwards in the buffer when in Sub state to see if
we saw an IAC in the prev value, add a new state: SubIac, that's entered
when we're in Sub and see Iac. If were in SubIac and see SE then we're
done. If we're in SubIac and we see anything other than SE, go back to
Sub.