hack: test against og parser w/ some code dupe
As I refactor the core parsing logic I don't want to break anything. The
easiest way to ensure that without having lots of unit tests is to
temporarily duplicate the internal parser buffer and the processing
logic.

For every receive, write the data to two buffers. Process one with the
new logic and one with the old. Assert the generated events and buffer
state are identical between the two.

In combination with the fuzzer we'll have high confidence the rewrite is
equivalent to the old code.